add base directories #109
