+++ config/machines/gaya.json
@@ -6,11 +6,14 @@
     "launch_options":["-bind-to core"],
     "reframe_base_dir":"/data/home/cladellash/benchmarking/build/reframe",
     "reports_base_dir":"/data/home/cladellash/benchmarking/reports/",
+    "input_dataset_base_dir":"/data/scratch/cladellash/input_data",
+    "output_app_dir":"/data/scratch/cladellash/feelppdb",
     "containers":[
         {
             "platform":"apptainer",
             "cachedir": "/home/u2/cladellash/.apptainer/cache",
-            "tmpdir": "/data/scratch/cladellash/images/tmp/"
+            "tmpdir": "/data/scratch/cladellash/images/tmp/",
+            "image_base_dir":"/data/scratch/cladellash/images/"
         }
     ]
 }
